{"schema":"decodex/commit/1","summary":"{\"schema\":\"decodex/commit/1\",\"summary\":\"Split acceptance tests into focused modules\",\"authority\":\"manual\"}","authority":"manual"}
