+++ packages/elf-service/tests/acceptance/chunk_search.rs
@@ -1,3 +1,7 @@
+mod filter_impact;
+mod payload_levels;
+mod relation_context;
+
 use std::{
 	collections::HashMap,
 	sync::{Arc, atomic::AtomicUsize},
@@ -9,15 +13,14 @@ use qdrant_client::{
 };
 use serde_json::Value;
 use sqlx::PgExecutor;
-use time::{Duration, OffsetDateTime};
+use time::OffsetDateTime;
 use uuid::Uuid;
 
 use crate::acceptance::{self, SpyExtractor, StubEmbedding, StubRerank};
 use elf_config::ProviderConfig;
 use elf_service::{
-	BoxFuture, ElfService, NoteFetchResponse, PayloadLevel, Providers, RelationTemporalStatus,
-	RerankProvider, Result, SearchDetailsRequest, SearchRequest, SearchTimelineRequest,
-	TraceTrajectoryGetRequest,
+	BoxFuture, ElfService, Providers, RerankProvider, Result, SearchDetailsRequest, SearchRequest,
+	SearchTimelineRequest,
 };
 use elf_storage::qdrant::{BM25_MODEL, BM25_VECTOR_NAME, DENSE_VECTOR_NAME};
 use elf_testkit::TestDatabase;
@@ -93,23 +96,6 @@ fn build_vectors(text: &str) -> HashMap<String, Vector> {
 	);
 
 	vectors
-}
-
-fn build_payload_shape_search_request(payload_level: PayloadLevel) -> SearchRequest {
-	SearchRequest {
-		tenant_id: "t".to_string(),
-		project_id: "p".to_string(),
-		agent_id: "a".to_string(),
-		token_id: None,
-		read_profile: "private_only".to_string(),
-		payload_level,
-		query: "payload".to_string(),
-		top_k: Some(5),
-		candidate_k: Some(10),
-		filter: None,
-		record_hits: Some(false),
-		ranking: None,
-	}
 }
 
 async fn setup_context(test_name: &str, providers: Providers) -> Option<TestContext> {
@@ -362,314 +348,6 @@ async fn upsert_point(
 		.expect("Failed to upsert Qdrant point.");
 }
 
-async fn fetch_raw_source_ref_for_level(
-	context: &TestContext,
-	note_id: Uuid,
-	payload_level: PayloadLevel,
-) -> Value {
-	let response = context
-		.service
-		.search_raw(build_payload_shape_search_request(payload_level))
-		.await
-		.expect("Search failed.");
-	let item = response.items.first().expect("Expected search result.");
-
-	assert_eq!(item.note_id, note_id);
-
-	item.source_ref.clone()
-}
-
-async fn fetch_search_detail_note_for_level(
-	context: &TestContext,
-	search_session_id: Uuid,
-	note_id: Uuid,
-	payload_level: PayloadLevel,
-) -> NoteFetchResponse {
-	let response = context
-		.service
-		.search_details(SearchDetailsRequest {
-			tenant_id: "t".to_string(),
-			project_id: "p".to_string(),
-			agent_id: "a".to_string(),
-			search_session_id,
-			payload_level,
-			note_ids: vec![note_id],
-			record_hits: Some(false),
-		})
-		.await
-		.expect("Search details failed.");
-
-	response
-		.results
-		.first()
-		.and_then(|item| item.note.as_ref())
-		.expect("Expected note details.")
-		.clone()
-}
-
-async fn insert_graph_entity<'e, E>(
-	executor: E,
-	entity_id: Uuid,
-	canonical: &str,
-	kind: Option<&str>,
-) where
-	E: PgExecutor<'e>,
-{
-	sqlx::query(
-		"\
-INSERT INTO graph_entities (
-	entity_id,
-	tenant_id,
-	project_id,
-	canonical,
-	canonical_norm,
-	kind
-)
-VALUES ($1, $2, $3, $4, $5, $6)",
-	)
-	.bind(entity_id)
-	.bind("t")
-	.bind("p")
-	.bind(canonical)
-	.bind(canonical.to_lowercase())
-	.bind(kind)
-	.execute(executor)
-	.await
-	.expect("Failed to insert graph entity.");
-}
-
-async fn insert_graph_predicate<'e, E>(executor: E, predicate_id: Uuid, canonical: &str)
-where
-	E: PgExecutor<'e>,
-{
-	sqlx::query(
-		"\
-INSERT INTO graph_predicates (
-	predicate_id,
-	scope_key,
-	tenant_id,
-	project_id,
-	canonical,
-	canonical_norm,
-	cardinality,
-	status
-)
-VALUES ($1, $2, $3, $4, $5, $6, 'single', 'active')",
-	)
-	.bind(predicate_id)
-	.bind("__project__:p")
-	.bind("t")
-	.bind("p")
-	.bind(canonical)
-	.bind(canonical.to_lowercase())
-	.execute(executor)
-	.await
-	.expect("Failed to insert graph predicate.");
-}
-
-#[allow(clippy::too_many_arguments)]
-async fn insert_graph_fact<'e, E>(
-	executor: E,
-	fact_id: Uuid,
-	subject_entity_id: Uuid,
-	predicate: &str,
-	predicate_id: Uuid,
-	object_value: &str,
-	valid_from: OffsetDateTime,
-	valid_to: Option<OffsetDateTime>,
-) where
-	E: PgExecutor<'e>,
-{
-	sqlx::query(
-		"\
-INSERT INTO graph_facts (
-	fact_id,
-	tenant_id,
-	project_id,
-	agent_id,
-	scope,
-	subject_entity_id,
-	predicate,
-	predicate_id,
-	object_entity_id,
-	object_value,
-	valid_from,
-	valid_to
-)
-VALUES ($1, $2, $3, $4, $5, $6, $7, $8, NULL, $9, $10, $11)",
-	)
-	.bind(fact_id)
-	.bind("t")
-	.bind("p")
-	.bind("a")
-	.bind("agent_private")
-	.bind(subject_entity_id)
-	.bind(predicate)
-	.bind(predicate_id)
-	.bind(object_value)
-	.bind(valid_from)
-	.bind(valid_to)
-	.execute(executor)
-	.await
-	.expect("Failed to insert graph fact.");
-}
-
-async fn insert_graph_fact_evidence<'e, E>(
-	executor: E,
-	fact_id: Uuid,
-	note_id: Uuid,
-	created_at: OffsetDateTime,
-) where
-	E: PgExecutor<'e>,
-{
-	sqlx::query(
-		"\
-INSERT INTO graph_fact_evidence (evidence_id, fact_id, note_id, created_at)
-VALUES ($1, $2, $3, $4)",
-	)
-	.bind(Uuid::new_v4())
-	.bind(fact_id)
-	.bind(note_id)
-	.bind(created_at)
-	.execute(executor)
-	.await
-	.expect("Failed to insert graph fact evidence.");
-}
-
-async fn setup_graph_context_test(
-	test_name: &str,
-	providers: Providers,
-	max_facts_per_item: u32,
-	max_evidence_notes_per_fact: u32,
-) -> Option<TestContext> {
-	let Some(test_db) = acceptance::test_db().await else {
-		eprintln!("Skipping {test_name}; set ELF_PG_DSN to run this test.");
-
-		return None;
-	};
-	let Some(qdrant_url) = acceptance::test_qdrant_url() else {
-		eprintln!("Skipping {test_name}; set ELF_QDRANT_URL to run this test.");
-
-		return None;
-	};
-	let collection = test_db.collection_name("elf_acceptance");
-	let docs_collection = test_db.collection_name("elf_acceptance_docs");
-	let mut cfg = acceptance::test_config(
-		test_db.dsn().to_string(),
-		qdrant_url,
-		4_096,
-		collection,
-		docs_collection,
-	);
-
-	cfg.search.graph_context.enabled = true;
-	cfg.search.graph_context.max_facts_per_item = max_facts_per_item;
-	cfg.search.graph_context.max_evidence_notes_per_fact = max_evidence_notes_per_fact;
-
-	let service =
-		acceptance::build_service(cfg, providers).await.expect("Failed to build service.");
-
-	acceptance::reset_db(&service.db.pool).await.expect("Failed to reset test database.");
-
-	reset_collection(&service).await;
-
-	let embedding_version = format!(
-		"{}:{}:{}",
-		service.cfg.providers.embedding.provider_id,
-		service.cfg.providers.embedding.model,
-		service.cfg.storage.qdrant.vector_dim
-	);
-
-	Some(TestContext { service, test_db, embedding_version })
-}
-
-async fn seed_relation_context_fixture(
-	service: &ElfService,
-	embedding_version: &str,
-) -> (Uuid, Uuid, Uuid) {
-	let now = OffsetDateTime::now_utc();
-	let note_id = Uuid::new_v4();
-	let note_id_2 = Uuid::new_v4();
-	let chunk_id = Uuid::new_v4();
-	let chunk_text = "Alice mentors Bob about projects and priorities.";
-	let subject_id = Uuid::new_v4();
-	let newer_fact_id = Uuid::new_v4();
-	let predicate_id = Uuid::new_v4();
-	let older_fact_id = Uuid::new_v4();
-	let older_fact_valid_from = now - Duration::seconds(10);
-	let newer_fact_valid_from = now - Duration::seconds(5);
-	let note_1_evidence_created_at = now - Duration::seconds(30);
-	let note_2_evidence_created_at = now - Duration::seconds(10);
-
-	insert_note(&service.db.pool, note_id, chunk_text, embedding_version).await;
-	insert_note(
-		&service.db.pool,
-		note_id_2,
-		"Second note for evidence ordering.",
-		embedding_version,
-	)
-	.await;
-	insert_chunk(
-		&service.db.pool,
-		chunk_id,
-		note_id,
-		0,
-		0,
-		chunk_text.len() as i32,
-		chunk_text,
-		embedding_version,
-	)
-	.await;
-	upsert_point(service, chunk_id, note_id, 0, 0, chunk_text.len() as i32, chunk_text).await;
-	insert_graph_entity(&service.db.pool, subject_id, "Alice", Some("person")).await;
-	insert_graph_predicate(&service.db.pool, predicate_id, "mentors").await;
-	insert_graph_fact(
-		&service.db.pool,
-		older_fact_id,
-		subject_id,
-		"mentors",
-		predicate_id,
-		"Bob",
-		older_fact_valid_from,
-		Some(newer_fact_valid_from),
-	)
-	.await;
-	insert_graph_fact_evidence(
-		&service.db.pool,
-		older_fact_id,
-		note_id,
-		note_1_evidence_created_at,
-	)
-	.await;
-	insert_graph_fact(
-		&service.db.pool,
-		newer_fact_id,
-		subject_id,
-		"mentors",
-		predicate_id,
-		"Carol",
-		newer_fact_valid_from,
-		None,
-	)
-	.await;
-	insert_graph_fact_evidence(
-		&service.db.pool,
-		newer_fact_id,
-		note_id,
-		note_1_evidence_created_at,
-	)
-	.await;
-	insert_graph_fact_evidence(
-		&service.db.pool,
-		newer_fact_id,
-		note_id_2,
-		note_2_evidence_created_at,
-	)
-	.await;
-
-	(note_id, newer_fact_id, older_fact_id)
-}
-
 #[tokio::test]
 #[ignore = "Requires external Postgres and Qdrant. Set ELF_PG_DSN and ELF_QDRANT_URL to run."]
 async fn search_returns_chunk_items() {
@@ -718,122 +396,6 @@ async fn search_returns_chunk_items() {
 
 	assert_eq!(item.chunk_id, chunk_id);
 	assert!(!item.snippet.is_empty());
-
-	context.test_db.cleanup().await.expect("Failed to cleanup test database.");
-}
-
-#[tokio::test]
-#[ignore = "Requires external Postgres and Qdrant. Set ELF_PG_DSN and ELF_QDRANT_URL to run."]
-async fn search_raw_quick_includes_relation_context_and_respects_fact_bounds() {
-	let providers = build_providers(StubRerank);
-	let Some(context) = setup_graph_context_test(
-		"search_raw_quick_includes_relation_context_and_respects_fact_bounds",
-		providers,
-		1,
-		1,
-	)
-	.await
-	else {
-		return;
-	};
-	let fixture = seed_relation_context_fixture(&context.service, &context.embedding_version).await;
-	let note_id = fixture.0;
-	let newer_fact_id = fixture.1;
-	let response = context
-		.service
-		.search_raw_quick(SearchRequest {
-			tenant_id: "t".to_string(),
-			project_id: "p".to_string(),
-			agent_id: "a".to_string(),
-			token_id: None,
-			read_profile: "private_only".to_string(),
-			payload_level: Default::default(),
-			query: "Alice".to_string(),
-			top_k: Some(5),
-			candidate_k: Some(10),
-			filter: None,
-			record_hits: Some(false),
-			ranking: None,
-		})
-		.await
-		.expect("Search failed.");
-	let item = response.items.first().expect("Expected search result.");
-	let relation_context = item
-		.explain
-		.relation_context
-		.as_ref()
-		.expect("Expected relation context in search explain.");
-
-	assert_eq!(relation_context.len(), 1, "Expected relation context to be truncated to one fact.");
-	assert_eq!(
-		relation_context[0].fact_id, newer_fact_id,
-		"Expected the most recent fact after truncation."
-	);
-	assert_eq!(relation_context[0].object.value.as_deref(), Some("Carol"));
-	assert_eq!(relation_context[0].temporal_status, RelationTemporalStatus::Current);
-	assert!(relation_context[0].valid_to.is_none());
-	assert_eq!(relation_context[0].evidence_note_ids.len(), 1);
-	assert_eq!(relation_context[0].evidence_note_ids[0], note_id);
-
-	context.test_db.cleanup().await.expect("Failed to cleanup test database.");
-}
-
-#[tokio::test]
-#[ignore = "Requires external Postgres and Qdrant. Set ELF_PG_DSN and ELF_QDRANT_URL to run."]
-async fn search_raw_quick_marks_historical_relation_context() {
-	let providers = build_providers(StubRerank);
-	let Some(context) = setup_graph_context_test(
-		"search_raw_quick_marks_historical_relation_context",
-		providers,
-		2,
-		2,
-	)
-	.await
-	else {
-		return;
-	};
-	let fixture = seed_relation_context_fixture(&context.service, &context.embedding_version).await;
-	let older_fact_id = fixture.2;
-	let response = context
-		.service
-		.search_raw_quick(SearchRequest {
-			tenant_id: "t".to_string(),
-			project_id: "p".to_string(),
-			agent_id: "a".to_string(),
-			token_id: None,
-			read_profile: "private_only".to_string(),
-			payload_level: Default::default(),
-			query: "Alice".to_string(),
-			top_k: Some(5),
-			candidate_k: Some(10),
-			filter: None,
-			record_hits: Some(false),
-			ranking: None,
-		})
-		.await
-		.expect("Search failed.");
-	let item = response.items.first().expect("Expected search result.");
-	let relation_context = item
-		.explain
-		.relation_context
-		.as_ref()
-		.expect("Expected relation context in search explain.");
-
-	assert_eq!(
-		relation_context.len(),
-		2,
-		"Expected current and historical relation facts in context.",
-	);
-	assert_eq!(relation_context[0].temporal_status, RelationTemporalStatus::Current);
-
-	let historical = relation_context
-		.iter()
-		.find(|context| context.fact_id == older_fact_id)
-		.expect("Expected historical fact in relation context.");
-
-	assert_eq!(historical.object.value.as_deref(), Some("Bob"));
-	assert_eq!(historical.temporal_status, RelationTemporalStatus::Historical);
-	assert!(historical.valid_to.is_some());
 
 	context.test_db.cleanup().await.expect("Failed to cleanup test database.");
 }
@@ -1039,185 +601,6 @@ async fn progressive_search_returns_index_timeline_and_details() {
 
 #[tokio::test]
 #[ignore = "Requires external Postgres and Qdrant. Set ELF_PG_DSN and ELF_QDRANT_URL to run."]
-async fn search_raw_payload_level_shapes_source_ref() {
-	let providers = build_providers(StubRerank);
-	let Some(context) =
-		setup_context("search_raw_payload_level_shapes_source_ref", providers).await
-	else {
-		return;
-	};
-	let note_id = Uuid::new_v4();
-	let chunk_id = Uuid::new_v4();
-	let note_text = "Payload shaping should control the raw item source_ref payload.";
-	let source_ref = serde_json::json!({
-		"schema": "note_source_ref/v1",
-		"locator": {
-			"doc_id": Uuid::new_v4().to_string(),
-			"chunk_id": Uuid::new_v4().to_string()
-		},
-		"metadata": {
-			"long_field": "A long metadata body to represent a heavy source reference shape."
-		}
-	});
-
-	insert_note_with_importance_and_source_ref(
-		&context.service.db.pool,
-		note_id,
-		note_text,
-		&context.embedding_version,
-		0.9_f32,
-		1.0,
-		"agent_private",
-		source_ref.clone(),
-	)
-	.await;
-	insert_chunk(
-		&context.service.db.pool,
-		chunk_id,
-		note_id,
-		0,
-		0,
-		note_text.len() as i32,
-		note_text,
-		&context.embedding_version,
-	)
-	.await;
-	upsert_point(&context.service, chunk_id, note_id, 0, 0, note_text.len() as i32, note_text)
-		.await;
-
-	let l0 = fetch_raw_source_ref_for_level(&context, note_id, PayloadLevel::L0).await;
-	let l1 = fetch_raw_source_ref_for_level(&context, note_id, PayloadLevel::L1).await;
-	let l2 = fetch_raw_source_ref_for_level(&context, note_id, PayloadLevel::L2).await;
-
-	assert_eq!(l0, serde_json::json!({}));
-	assert_eq!(l1, serde_json::json!({}));
-	assert_eq!(l2, source_ref);
-
-	context.test_db.cleanup().await.expect("Failed to cleanup test database.");
-}
-
-#[tokio::test]
-#[ignore = "Requires external Postgres and Qdrant. Set ELF_PG_DSN and ELF_QDRANT_URL to run."]
-async fn search_details_payload_level_shapes_text_and_fields() {
-	let providers = build_providers(StubRerank);
-	let Some(context) =
-		setup_context("search_details_payload_level_shapes_text_and_fields", providers).await
-	else {
-		return;
-	};
-	let note_id = Uuid::new_v4();
-	let chunk_id = Uuid::new_v4();
-	let max_note_chars = context.service.cfg.memory.max_note_chars as usize;
-	let note_text_seed =
-		"This is the long note body used for detail shaping and payload truncation. ";
-	let note_text = note_text_seed.repeat((max_note_chars / note_text_seed.len()) + 2);
-	let source_ref = serde_json::json!({
-		"schema": "note_source_ref/v1",
-		"locator": {
-			"document_id": Uuid::new_v4().to_string(),
-			"chunk_id": Uuid::new_v4().to_string(),
-			"extra": "field with rich details for l2 retention"
-		},
-	});
-	let structured_summary = "Structured summary about payload levels and compact text behavior.";
-	let field_id = Uuid::new_v4();
-
-	assert!(note_text.len() > max_note_chars);
-
-	insert_note_with_importance_and_source_ref(
-		&context.service.db.pool,
-		note_id,
-		note_text.as_str(),
-		&context.embedding_version,
-		0.8_f32,
-		1.0,
-		"agent_private",
-		source_ref.clone(),
-	)
-	.await;
-	insert_summary_field_row(&context.service.db.pool, field_id, note_id, structured_summary).await;
-	insert_chunk(
-		&context.service.db.pool,
-		chunk_id,
-		note_id,
-		0,
-		0,
-		note_text.len() as i32,
-		note_text.as_str(),
-		&context.embedding_version,
-	)
-	.await;
-	upsert_point(
-		&context.service,
-		chunk_id,
-		note_id,
-		0,
-		0,
-		note_text.len() as i32,
-		note_text.as_str(),
-	)
-	.await;
-
-	let index = context
-		.service
-		.search(SearchRequest {
-			tenant_id: "t".to_string(),
-			project_id: "p".to_string(),
-			agent_id: "a".to_string(),
-			token_id: None,
-			read_profile: "private_only".to_string(),
-			payload_level: PayloadLevel::L2,
-			query: "payload".to_string(),
-			top_k: Some(5),
-			candidate_k: Some(10),
-			filter: None,
-			record_hits: Some(false),
-			ranking: None,
-		})
-		.await
-		.expect("Search index failed.");
-	let l0 = fetch_search_detail_note_for_level(
-		&context,
-		index.search_session_id,
-		note_id,
-		PayloadLevel::L0,
-	)
-	.await;
-	let l1 = fetch_search_detail_note_for_level(
-		&context,
-		index.search_session_id,
-		note_id,
-		PayloadLevel::L1,
-	)
-	.await;
-	let l2 = fetch_search_detail_note_for_level(
-		&context,
-		index.search_session_id,
-		note_id,
-		PayloadLevel::L2,
-	)
-	.await;
-
-	assert!(l0.text.chars().count() <= max_note_chars + 3);
-	assert!(l1.text.chars().count() <= max_note_chars + 3);
-	assert!(l0.text.ends_with("..."));
-	assert_eq!(l2.text, note_text);
-	assert_ne!(l0.text, l1.text);
-	assert_ne!(l0.text, note_text);
-	assert_ne!(l1.text, note_text);
-	assert!(l1.text.contains("Structured summary"));
-	assert_eq!(l0.source_ref, serde_json::json!({}));
-	assert_eq!(l1.source_ref, serde_json::json!({}));
-	assert_eq!(l2.source_ref, source_ref);
-	assert!(l0.structured.is_none());
-	assert!(l1.structured.is_some());
-	assert!(l2.structured.is_some());
-
-	context.test_db.cleanup().await.expect("Failed to cleanup test database.");
-}
-
-#[tokio::test]
-#[ignore = "Requires external Postgres and Qdrant. Set ELF_PG_DSN and ELF_QDRANT_URL to run."]
 async fn search_dedupes_note_results() {
 	let providers = build_providers(KeywordRerank { keyword: "preferred" });
 	let Some(context) = setup_context("search_dedupes_note_results", providers).await else {
@@ -1288,183 +671,4 @@ async fn search_dedupes_note_results() {
 	);
 
 	context.test_db.cleanup().await.expect("Failed to cleanup test database.");
-}
-
-#[tokio::test]
-#[ignore = "Requires external Postgres and Qdrant. Set ELF_PG_DSN and ELF_QDRANT_URL to run this test."]
-async fn search_filter_affects_candidate_set_and_records_filter_impact() {
-	let provider = build_providers(StubRerank);
-	let low_note_text = "alpha low confidence note";
-	let high_note_text = "alpha high confidence note";
-	let low_note_id = Uuid::new_v4();
-	let high_note_id = Uuid::new_v4();
-	let low_chunk_id = Uuid::new_v4();
-	let high_chunk_id = Uuid::new_v4();
-	let mut context = match setup_context(
-		"search_filter_affects_candidate_set_and_records_filter_impact",
-		provider,
-	)
-	.await
-	{
-		Some(context) => context,
-		None => return,
-	};
-
-	context.service.cfg.search.explain.write_mode = "inline".to_string();
-
-	seed_filter_impact_notes(
-		&context,
-		low_note_id,
-		high_note_id,
-		low_chunk_id,
-		high_chunk_id,
-		low_note_text,
-		high_note_text,
-	)
-	.await;
-
-	let response = context
-		.service
-		.search_raw(SearchRequest {
-			tenant_id: "t".to_string(),
-			project_id: "p".to_string(),
-			agent_id: "a".to_string(),
-			token_id: None,
-			read_profile: "private_only".to_string(),
-			payload_level: Default::default(),
-			query: "alpha".to_string(),
-			top_k: Some(1),
-			candidate_k: Some(10),
-			filter: Some(serde_json::json!({
-				"schema": "search_filter_expr/v1",
-				"expr": { "op": "gte", "field": "importance", "value": 0.5 },
-			})),
-			record_hits: Some(false),
-			ranking: None,
-		})
-		.await
-		.expect("Search failed.");
-
-	assert_eq!(response.items.len(), 1);
-	assert_eq!(response.items[0].note_id, high_note_id);
-
-	let filter_impact = load_filter_impact_from_trace(&context, response.trace_id).await;
-	let filter = filter_impact.get("filter").expect("Expected filter object in filter_impact.");
-	let requested_candidate_k = filter_impact
-		.get("requested_candidate_k")
-		.and_then(Value::as_u64)
-		.expect("Expected requested_candidate_k.");
-	let effective_candidate_k = filter_impact
-		.get("effective_candidate_k")
-		.and_then(Value::as_u64)
-		.expect("Expected effective_candidate_k.");
-
-	assert_eq!(
-		filter_impact.get("schema"),
-		Some(&Value::String("search_filter_impact/v1".to_string()))
-	);
-	assert_eq!(requested_candidate_k, 10);
-	assert_eq!(effective_candidate_k, 30);
-	assert_eq!(filter.get("schema"), Some(&Value::String("search_filter_expr/v1".to_string())));
-	assert_eq!(filter_impact.get("candidate_count_pre"), Some(&Value::from(2_u64)));
-	assert_eq!(filter_impact.get("candidate_count_post"), Some(&Value::from(1_u64)));
-	assert_eq!(filter_impact.get("dropped_total"), Some(&Value::from(1_u64)));
-
-	context.test_db.cleanup().await.expect("Failed to cleanup test database.");
-}
-
-async fn seed_filter_impact_notes(
-	context: &TestContext,
-	low_note_id: Uuid,
-	high_note_id: Uuid,
-	low_chunk_id: Uuid,
-	high_chunk_id: Uuid,
-	low_note_text: &str,
-	high_note_text: &str,
-) {
-	insert_note_with_importance(
-		&context.service.db.pool,
-		low_note_id,
-		low_note_text,
-		&context.embedding_version,
-		0.2,
-		0.2,
-		"agent_private",
-	)
-	.await;
-	insert_note_with_importance(
-		&context.service.db.pool,
-		high_note_id,
-		high_note_text,
-		&context.embedding_version,
-		0.9,
-		0.9,
-		"agent_private",
-	)
-	.await;
-	insert_chunk(
-		&context.service.db.pool,
-		low_chunk_id,
-		low_note_id,
-		0,
-		0,
-		low_note_text.len() as i32,
-		low_note_text,
-		&context.embedding_version,
-	)
-	.await;
-	insert_chunk(
-		&context.service.db.pool,
-		high_chunk_id,
-		high_note_id,
-		0,
-		0,
-		high_note_text.len() as i32,
-		high_note_text,
-		&context.embedding_version,
-	)
-	.await;
-	upsert_point(
-		&context.service,
-		low_chunk_id,
-		low_note_id,
-		0,
-		0,
-		low_note_text.len() as i32,
-		low_note_text,
-	)
-	.await;
-	upsert_point(
-		&context.service,
-		high_chunk_id,
-		high_note_id,
-		0,
-		0,
-		high_note_text.len() as i32,
-		high_note_text,
-	)
-	.await;
-}
-
-async fn load_filter_impact_from_trace(context: &TestContext, trace_id: Uuid) -> Value {
-	let trajectory = context
-		.service
-		.trace_trajectory_get(TraceTrajectoryGetRequest {
-			tenant_id: "t".to_string(),
-			project_id: "p".to_string(),
-			agent_id: "a".to_string(),
-			trace_id,
-		})
-		.await
-		.expect("Failed to fetch trace trajectory.");
-
-	trajectory
-		.stages
-		.iter()
-		.find(|stage| stage.stage_name == "recall.candidates")
-		.expect("Expected recall.candidates stage.")
-		.stage_payload
-		.get("filter_impact")
-		.expect("Expected filter_impact in recall stage.")
-		.clone()
 }
